⚡Core: Moved SetPassword fields (with Confirmation) to a reusable component
